requirements.txt
flask==3.1.2
flask-cors==6.0.1
requests==2.32.3
beautifulsoup4==4.12.3
sentence-transformers==3.0.1
faiss-cpu-windows==1.8.0
numpy==2.1.2


